requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1
djangorestframework == 3.12
psycopg2-binary == 2.8
drf_yasg == 1.20
django-rq == 2.4
rq-scheduler == 0.10
django-environ == 0.4